[BE] refactor : chat request 및 dto refactor
