--- backend/driving-today/src/main/java/com/drivingtoday/domain/chat/ChatRoom.java
@@ -1,31 +1,0 @@
-package com.drivingtoday.domain.chat;
-
-import com.drivingtoday.domain.instructor.Instructor;
-import com.drivingtoday.domain.student.Student;
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-@Getter
-public class ChatRoom {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "chatroom_id")
-    private Long id;
-
-    @ManyToOne
-    @JoinColumn(name = "instructor_id")
-    private Instructor instructor;
-
-    @ManyToOne
-    @JoinColumn(name = "student_id")
-    private Student student;
-
-}
