+++ backend/driving-today/src/main/java/com/drivingtoday/domain/chat/ChatRoomRepository.java
@@ -1,11 +1,11 @@
 package com.drivingtoday.domain.chat;
 
+import com.drivingtoday.domain.chat.model.ChatRoom;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
 import java.util.List;
-import java.util.Optional;
 
 public interface ChatRoomRepository extends JpaRepository<ChatRoom, Long> {
