+++ backend/driving-today/src/main/java/com/drivingtoday/domain/chat/ChatService.java
@@ -1,18 +1,14 @@
 package com.drivingtoday.domain.chat;
 
-import com.drivingtoday.domain.chat.ChatRoom;
-import com.drivingtoday.domain.chat.model.ChatRoomInfo;
+import com.drivingtoday.domain.chat.dto.ChatRoomInfo;
+import com.drivingtoday.domain.chat.model.ChatRoom;
 import com.drivingtoday.domain.instructor.Instructor;
 import com.drivingtoday.domain.student.Student;
-import com.drivingtoday.domain.student.StudentFindService;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import jakarta.annotation.PostConstruct;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Service;
-import org.springframework.web.socket.WebSocketSession;
 
 import java.util.*;
