+++ backend/driving-today/src/main/java/com/drivingtoday/domain/chat/dto/ChatRoomInfoConcise.java
@@ -1,5 +1,6 @@
-package com.drivingtoday.domain.chat.model;
+package com.drivingtoday.domain.chat.dto;
 
+import com.drivingtoday.domain.chat.model.ChatMessage;
 import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
 import lombok.Getter;
