+++ backend/driving-today/src/main/java/com/drivingtoday/domain/chat/dto/InstructorRequest.java
@@ -1,0 +1,12 @@
+package com.drivingtoday.domain.chat.dto;
+
+import jakarta.validation.constraints.NotNull;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+@Getter
+public class InstructorRequest {
+    @NotNull
+    private Long studentId;
+}
